Fixes group interactions with group specific terms
